+++ azsos_core/__init__.py
@@ -1,5 +1,5 @@
 """AZSOS core package."""
 
-__version__ = "0.3.4"
+__version__ = "0.3.5"
 FORMAT_VERSION = "azsos.v1"
 PACKAGE_EXTENSION = ".azsos"